requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
boto3==1.28.26
botocore==1.31.26
certifi==2023.7.22
charset-normalizer==3.2.0
discord.py==2.3.2
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
outcome==1.2.0
packaging==23.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.1
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==1.26.16
webdriver-manager==4.0.0
wsproto==1.2.0
yarl==1.9.2
